+++ src/domain/movieHandler.ts
@@ -2,22 +2,15 @@ import type { Movie } from "./../types/type";
 
 interface MovieHandler {
   movies: Movie[];
-  currentPageNumber: number;
   addMovies(movies: Movie[]): void;
-  sortByPopularity(movies: Movie[]): Movie[];
   initializeMovies(): void;
 }
 
 const movieHandler: MovieHandler = {
   movies: [],
-  currentPageNumber: 1,
 
   addMovies(movies) {
     this.movies = [...this.movies, ...movies];
-  },
-
-  sortByPopularity(movies) {
-    return movies.sort((a, b) => b.popularity - a.popularity);
   },
 
   initializeMovies() {
